feat(modules): add support for replying to messages with links only

+++ src/modules/plugins/ytdlp.py
@@ -7,26 +7,18 @@ from typing import ClassVar
 import orjson
 import regex as re
 from humanize import naturalsize
-from telethon.events import NewMessage
+from telethon.events import CallbackQuery, NewMessage
 from yt_dlp import YoutubeDL
 
 from src import PARENT_DIR, TMP_DIR
 from src.modules.base import ModuleBase
 from src.utils.command import Command
 from src.utils.downloads import upload_file
+from src.utils.filters import has_valid_url
 from src.utils.json import json_options, process_dict
+from src.utils.patterns import HTTP_URL_PATTERN, YOUTUBE_URL_PATTERN
 from src.utils.run import run_subprocess_shell
-from src.utils.telegram import edit_or_send_as_file
-
-YOUTUBE_URL_PATTERN = (
-    r'(?:https?:\/\/)?(?:www\.)?(?:youtube\.com|youtu\.be)'
-    r'\/(?:watch\?v=)?(?:embed\/)?(?:v\/)?(?:shorts\/)?(?:live\/)?'
-    r'(?:(?:watch\?)?(?:time_continue=(?:\d+))?\&?(?:v=))?([^\s&]+)'
-)
-HTTP_URL_PATTERN = (
-    r'https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}'
-    r'\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)'
-)
+from src.utils.telegram import edit_or_send_as_file, get_reply_message
 
 cookies_file = Path(PARENT_DIR) / 'cookies.txt'
 cookies = {'cookiefile': str(cookies_file.absolute())} if cookies_file.exists() else {}
@@ -36,23 +28,29 @@ params = {
 }
 
 
-async def get_youtube_info(event: NewMessage.Event) -> None:
+async def get_youtube_info(event: NewMessage.Event | CallbackQuery.Event) -> None:
     progress_message = await event.reply('Fetching video information...')
-    link = event.message.raw_text.split('ytinfo ')[1].strip()
+    message = (
+        await get_reply_message(event, previous=True)
+        if isinstance(event, CallbackQuery.Event)
+        else event.message
+    )
+    link = re.search(HTTP_URL_PATTERN, message.raw_text).group(0)
     try:
         info_dict = await get_running_loop().run_in_executor(
             None, partial(YoutubeDL(params).extract_info, link, download=False)
         )
         processed_info = process_dict(info_dict)
         json_str = orjson.dumps(processed_info, option=json_options).decode()
-        await edit_or_send_as_file(
+        edited = await edit_or_send_as_file(
             event,
             progress_message,
             text=f'<pre>{json_str}</pre>',
             file_name=f"{info_dict['id']}.json",
             caption=info_dict['webpage_url'],
         )
-        await progress_message.delete()
+        if not edited:
+            await progress_message.delete()
     except Exception as e:  # noqa: BLE001
         await progress_message.edit(f'An error occurred:\n<pre>{e!s}</pre>')
 
@@ -79,7 +77,16 @@ async def convert_subtitles(input_file: Path, srt_file: Path, txt_file: Path) ->
 
 async def get_youtube_subtitles(event: NewMessage.Event) -> None:
     progress_message = await event.reply('Downloading subtitles...')
-    language, link = event.message.raw_text.split('ytsub ')[1].split(maxsplit=2)
+    message = (
+        await get_reply_message(event, previous=True)
+        if isinstance(event, CallbackQuery.Event)
+        else event.message
+    )
+    link = re.search(HTTP_URL_PATTERN, message.raw_text).group(0)
+    if match := re.search(r'\s+([a-z]{{2}})\s+', message.raw_text):
+        language = match.group(1)
+    else:
+        language = 'ar'
     ydl_opts = {
         **params,
         'skip_download': True,
@@ -119,9 +126,14 @@ async def get_youtube_subtitles(event: NewMessage.Event) -> None:
     await progress_message.delete()
 
 
-async def get_youtube_formats(event: NewMessage.Event) -> None:
+async def get_youtube_formats(event: NewMessage.Event | CallbackQuery.Event) -> None:
     progress_message = await event.reply('Fetching available formats...')
-    link = event.message.raw_text.split('ytformats ')[1].strip()
+    message = (
+        await get_reply_message(event, previous=True)
+        if isinstance(event, CallbackQuery.Event)
+        else event.message
+    )
+    link = re.search(HTTP_URL_PATTERN, message.raw_text).group(0)
     try:
         ydl_opts = {**params, 'listformats': True}
         info_dict = await get_running_loop().run_in_executor(
@@ -132,16 +144,13 @@ async def get_youtube_formats(event: NewMessage.Event) -> None:
             await progress_message.edit('No formats found.')
             return
 
-        format_list = []
-        for f in formats:
-            format_id = f.get('format_id', 'N/A')
-            ext = f.get('ext', 'N/A')
-            resolution = f.get('resolution', 'N/A')
-            filesize = f.get('filesize', 'N/A')
-            if isinstance(filesize, int):
-                filesize = naturalsize(filesize, binary=True)
-            format_list.append(f'🆔 {format_id} | 📁 {ext} | 🖥️ {resolution} | 💾 {filesize}')
-
+        format_list = [
+            (
+                f'🖥 {f.get('format', 'N/A')} | 📁 {f.get('ext', 'N/A')} | '
+                f'💾 {naturalsize(f.get('filesize_approx', 0) or 0, binary=True)}'
+            )
+            for f in formats
+        ]
         await edit_or_send_as_file(
             event,
             progress_message,
@@ -161,17 +170,23 @@ class YTDLP(ModuleBase):
             handler=get_youtube_formats,
             description='[url]: Get available formats for a YouTube video.',
             pattern=re.compile(rf'^/ytformats\s+{HTTP_URL_PATTERN}$'),
+            condition=has_valid_url,
+            is_applicable_for_reply=True,
         ),
         'ytinfo': Command(
+            name='ytinfo',
             handler=get_youtube_info,
             description='[url]: Get video information as JSON.',
             pattern=re.compile(rf'^/ytinfo\s+{HTTP_URL_PATTERN}$'),
+            condition=has_valid_url,
+            is_applicable_for_reply=True,
         ),
         'ytsub': Command(
+            name='ytsub',
             handler=get_youtube_subtitles,
             description='[lang] [url]: Get YouTube video subtitles.',
             pattern=re.compile(rf'^/ytsub\s+([a-z]{{2}})\s+{YOUTUBE_URL_PATTERN}$'),
-            # condition=lambda event, message: True,  # Always applicable
-            # is_applicable_for_reply=True,
+            condition=has_valid_url,
+            is_applicable_for_reply=True,
         ),
     }
